Add "support_cus_res" option more graphics support
Added the "support_cus_res" option, made to support custom screen resolutions for the guest. Note that this option only takes effect when the VGA adapter is supported (virtio and qxl are supported, for example). Thanks to this option, better graphics support have been implemented (for supported VGA adapters ONLY).